+++ src/tsp_params.h
@@ -16,7 +16,7 @@ struct TSPSolverParams {
 struct TSPSearchParams {
   TSPSearchParams(const Nan::FunctionCallbackInfo<v8::Value>& info);
 
-  std::int64_t computeTimeLimit;
+  std::int32_t computeTimeLimit;
   std::int32_t depotNode;
 
   v8::Local<v8::Function> callback;
@@ -58,7 +58,7 @@ TSPSearchParams::TSPSearchParams(const Nan::FunctionCallbackInfo<v8::Value>& inf
   if (!computeTimeLimitOk || !depotNodeOk)
     throw std::runtime_error{"SearchOptions expects 'computeTimeLimit' (Number), 'depotNode' (Number)"};
 
-  computeTimeLimit = Nan::To<std::int64_t>(maybeComputeTimeLimit.ToLocalChecked()).FromJust();
+  computeTimeLimit = Nan::To<std::int32_t>(maybeComputeTimeLimit.ToLocalChecked()).FromJust();
   depotNode = Nan::To<std::int32_t>(maybeDepotNode.ToLocalChecked()).FromJust();
   callback = info[1].As<v8::Function>();
 }
